# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Project Intelligencia Scheduler API
  description: >
    API specification for the Smart Classroom & Timetable Scheduler platform.
    This is a comprehensive multi-tenant system with multiple microservices.
  version: "1.0.0"
  contact:
    name: Project Intelligencia Scheduler Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.intelligencia-scheduler.edu
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer

    # IAM Service Schemas
...
    InstitutionCreate:
      type: object
      required:
        - name
        - code
        - contact_email
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        contact_email:
          type: string
          format: email
        contact_phone:
          type: string
        website:
          type: string
        logo_url:
          type: string
    InstitutionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        address:
          type: string
        contact_email:
          type: string
          format: email
        contact_phone:
          type: string
        website:
          type: string
        logo_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DepartmentCreate:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        head_user_id:
          type: string
          format: uuid
    DepartmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        institution_id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        description:
          type: string
        head_user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ClassroomCreate:
      type: object
      required:
        - name
        - capacity
        - room_type_id
      properties:
        name:
          type: string
        building:
          type: string
        floor:
          type: string
        capacity:
          type: integer
        room_type_id:
          type: string
          format: uuid
        is_lab:
          type: boolean
          default: false
        has_projector:
          type: boolean
          default: false
        has_computers:
          type: boolean
          default: false
        additional_facilities:
          type: string
    ClassroomResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        institution_id:
          type: string
          format: uuid
        name:
          type: string
        building:
          type: string
        floor:
          type: string
        capacity:
          type: integer
        room_type_id:
          type: string
          format: uuid
        room_type:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        is_lab:
          type: boolean
        has_projector:
          type: boolean
        has_computers:
          type: boolean
        additional_facilities:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/institutions:
    get:
      tags:
        - Institutions
      summary: Get all institutions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/InstitutionResponse'
    
    post:
      tags:
        - Institutions
      summary: Create a new institution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionResponse'
  /api/v1/institutions/{id}:
    get:
      tags:
        - Institutions
      summary: Get institution by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionResponse'
        '404':
          description: Institution not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Institutions
      summary: Update an institution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionCreate'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionResponse'
        '404':
          description: Institution not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Institutions
      summary: Delete an institution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Institution not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/departments:
    get:
      tags:
        - Departments
      summary: Get all departments
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/DepartmentResponse'
    
    post:
      tags:
        - Departments
      summary: Create a new department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
  /api/v1/departments/{id}:
    get:
      tags:
        - Departments
      summary: Get department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Departments
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreate'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Departments
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Similar patterns repeated for classrooms, subjects, faculty, batches, etc.
  # Only showing a few examples to keep the file manageable
  /api/v1/classrooms:
    get:
      tags:
        - Classrooms
      summary: Get all classrooms
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
        - name: room_type_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_lab
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClassroomResponse'
    
    post:
      tags:
        - Classrooms
      summary: Create a new classroom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassroomCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassroomResponse'
...